sync-engine/requirements.txt
MetaTrader5>=5.0.45
mt5linux>=0.1.8
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
redis>=5.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
cryptography>=42.0.0
pydantic>=2.5.0
httpx>=0.26.0
pyjwt>=2.8.0
